Strip Riak Search hook
If Riak Search is not enabled then make sure to stip its pre-commit
hook out of all buckets.  Because of a bug in previous versions of
Riak (with `riak_core_bucket:set_bucket`) bucket fixups can leak
changes into the raw ring.  This means that when Riak Search is
disabled its pre-commit hook can appear in the list EVEN THOUGH the
`search` property is set to `false`.  This is because Riak Search
doesn't register bucket fixups if it is not enabled.

This patch checks if RS is enabled.  If it isn't it makes sure that
the Riak Search pre-commit hook is stripped from all buckets.
Preventing runtime errors.