WIP: Remove "flag transition" code
These steps not be taken on behalf of the user. They can be expensive
or undesired. If the user associates a type/bucket with existing data
then a command should be run manually to fold and index the data. If
the user disassociates an index then the user should run a delete by
query for that type/bucket IFF the user wants to.